Reward model moved to player module
