+++ src/Player/Reward.php
@@ -1,7 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Game\Dungeon;
+namespace Game\Player;
+
+use Game\Dungeon\Drop;
 
 class Reward
 {
